backend/requirements.txt
# FastAPI & Server
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Database
pymongo==4.6.3
motor==3.3.2

# ML / AI - Emotion Detection
transformers==4.39.3
torch>=2.6.0
torchvision>=0.21.0
Pillow>=10.4.0
numpy>=1.26.4
scikit-learn>=1.4.0

# Audio Processing
librosa>=0.10.1
soundfile>=0.12.1

# Image Processing
opencv-python-headless>=4.9.0.80

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# AI API for Recommendations
openai>=1.12.0

# Utilities
python-dotenv>=1.0.1
pydantic>=2.6.4
pydantic[email]>=2.6.4
pydantic-settings>=2.2.1
httpx>=0.27.0
pydub==0.25.1
deepface>=0.0.95
tf_keras>=2.15
